depthCamera: moved definition of private types into a custom header file
