+++ src/devices/depthCamera/CMakeLists.txt
@@ -37,7 +37,9 @@ if(ENABLE_depthCamera)
                       ${YARP_sig_INCLUDE_DIRS}
                       ${YARP_dev_INCLUDE_DIRS})
 
-  yarp_add_plugin(yarp_depthCamera depthCameraDriver.cpp depthCameraDriver.h)
+  yarp_add_plugin(yarp_depthCamera depthCameraDriver.cpp depthCameraDriver.h depthCameraImpl.hpp)
+  
+  set_property(TARGET yarp_depthCamera PROPERTY PRIVATE_HEADER depthCameraImpl.hpp)
 
   target_link_libraries(yarp_depthCamera YARP_OS YARP_sig YARP_dev ${OPENNI2_LIBRARY})
